itinerary-app/requirements.txt
# dependencies for weather microservice
fastapi
uvicorn
python-dotenv
requests
pydantic
pytest
pytest-mock
httpx